+++ scenedetect/stats_manager.py
@@ -118,41 +118,35 @@ class StatsManager:
         """Register a list of metric keys that will be used by the detector."""
         self._metric_keys = self._metric_keys.union(set(metric_keys))
 
-    # TODO(v1.0): Change frame_number to a FrameTimecode now that it is just a hash and will
-    # be required for VFR support. This API is also really difficult to use, this type should just
-    # function like a dictionary.
-    def get_metrics(self, frame_number: int, metric_keys: ty.Iterable[str]) -> ty.List[ty.Any]:
-        """Return the requested statistics/metrics for a given frame.
-
-        Arguments:
-            frame_number (int): Frame number to retrieve metrics for.
-            metric_keys (List[str]): A list of metric keys to look up.
+    # TODO(v1.0): This interface is difficult to use, we should support the dictionary protocol.
+    def get_metrics(
+        self, timecode: FrameTimecode, metric_keys: ty.Iterable[str]
+    ) -> ty.List[ty.Any]:
+        """Return the requested statistics/metrics for a given timecode.
 
         Returns:
-            A list containing the requested frame metrics for the given frame number
-            in the same order as the input list of metric keys. If a metric could
-            not be found, None is returned for that particular metric.
+            A list containing the requested frame metrics for the given frame number, ordered as
+            they are in `metric_keys`.
         """
-        return [self._get_metric(frame_number, metric_key) for metric_key in metric_keys]
+        return [self._get_metric(timecode, metric_key) for metric_key in metric_keys]
 
-    def set_metrics(self, frame_number: int, metric_kv_dict: ty.Dict[str, ty.Any]) -> None:
+    def set_metrics(self, timecode: FrameTimecode, metric_kv_dict: ty.Dict[str, ty.Any]) -> None:
         """Set Metrics: Sets the provided statistics/metrics for a given frame.
 
         Arguments:
-            frame_number: Frame number to retrieve metrics for.
-            metric_kv_dict: A dict mapping metric keys to the
-                respective integer/floating-point metric values to set.
+            timecode: Timecode to set metrics for.
+            metric_kv_dict: Key value mapping of metrics to their values for `timecode`.
         """
         for metric_key in metric_kv_dict:
-            self._set_metric(frame_number, metric_key, metric_kv_dict[metric_key])
+            self._set_metric(timecode, metric_key, metric_kv_dict[metric_key])
 
-    def metrics_exist(self, frame_number: int, metric_keys: ty.Iterable[str]) -> bool:
+    def metrics_exist(self, timecode: FrameTimecode, metric_keys: ty.Iterable[str]) -> bool:
         """Metrics Exist: Checks if the given metrics/stats exist for the given frame.
 
         Returns:
             bool: True if the given metric keys exist for the frame, False otherwise.
         """
-        return all([self._metric_exists(frame_number, metric_key) for metric_key in metric_keys])
+        return all([self._metric_exists(timecode, metric_key) for metric_key in metric_keys])
 
     def is_save_required(self) -> bool:
         """Is Save Required: Checks if the stats have been updated since loading.
@@ -166,23 +160,17 @@ class StatsManager:
     def save_to_csv(
         self,
         csv_file: ty.Union[str, bytes, Path, ty.TextIO],
-        base_timecode: ty.Optional[FrameTimecode] = None,
         force_save=True,
     ) -> None:
         """Save To CSV: Saves all frame metrics stored in the StatsManager to a CSV file.
 
         Arguments:
             csv_file: A file handle opened in write mode (e.g. open('...', 'w')) or a path as str.
-            base_timecode: [DEPRECATED] DO NOT USE. For backwards compatibility.
             force_save: If True, writes metrics out even if an update is not required.
 
         Raises:
             OSError: If `path` cannot be opened or a write failure occurs.
         """
-        # TODO(v0.7): Replace with DeprecationWarning that `base_timecode` will be removed in v0.8.
-        if base_timecode is not None:
-            logger.error("base_timecode is deprecated and has no effect.")
-
         if not (force_save or self.is_save_required()):
             logger.info("No metrics to write.")
             return
@@ -200,9 +188,8 @@ class StatsManager:
         frame_keys = sorted(self._frame_metrics.keys())
         logger.info("Writing %d frames to CSV...", len(frame_keys))
         for frame_key in frame_keys:
-            frame_timecode = self._base_timecode + frame_key
             csv_writer.writerow(
-                [frame_timecode.get_frames() + 1, frame_timecode.get_timecode()]
+                [frame_key.get_frames() + 1, frame_key.get_timecode()]
                 + [str(metric) for metric in self.get_metrics(frame_key, metric_keys)]
             )
 
@@ -303,18 +290,16 @@ class StatsManager:
 
     # TODO: Get rid of these functions and simplify the implementation of this class.
 
-    def _get_metric(self, frame_number: int, metric_key: str) -> ty.Optional[ty.Any]:
-        if self._metric_exists(frame_number, metric_key):
-            return self._frame_metrics[frame_number][metric_key]
+    def _get_metric(self, timecode: FrameTimecode, metric_key: str) -> ty.Optional[ty.Any]:
+        if self._metric_exists(timecode, metric_key):
+            return self._frame_metrics[timecode][metric_key]
         return None
 
-    def _set_metric(self, frame_number: int, metric_key: str, metric_value: ty.Any) -> None:
+    def _set_metric(self, timecode: FrameTimecode, metric_key: str, metric_value: ty.Any) -> None:
         self._metrics_updated = True
-        if frame_number not in self._frame_metrics:
-            self._frame_metrics[frame_number] = dict()
-        self._frame_metrics[frame_number][metric_key] = metric_value
+        if timecode not in self._frame_metrics:
+            self._frame_metrics[timecode] = dict()
+        self._frame_metrics[timecode][metric_key] = metric_value
 
-    def _metric_exists(self, frame_number: int, metric_key: str) -> bool:
-        return (
-            frame_number in self._frame_metrics and metric_key in self._frame_metrics[frame_number]
-        )
+    def _metric_exists(self, timecode: FrameTimecode, metric_key: str) -> bool:
+        return timecode in self._frame_metrics and metric_key in self._frame_metrics[timecode]